core: use CypherOS version to signal mIsUpgrade
M: We use a static fingerprint that is only changed when a new OEM build
is released, so every flash shows Android is starting instead of
upgrading.  This will fix that.

N: even though we dont have the starting/upgrading screen on N yet, this
is still needed to delete the correct caches, and grant/deny specific
runtime permissions like a true oem update would do

Updated for Nougat By: @beanstown106